Refactoring:   - Rename louis.h to the more descriptive internal.h.   - Add utils.c for common utility functions.     This can later be extended further. re #318
