+++ liblouis/internal.h
