+++ liblouis/wrappers.c
@@ -27,7 +27,7 @@ License along with liblouis. If not, see <http://www.gnu.org/licenses/>.
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "louis.h"
+#include "internal.h"
 
 static int
 ignoreCaseComp (const char *str1, const char *str2, size_t length)
